Centralize definition of brand identifiers
Prior to this commit, the name of the Automation Voice and its "vendor"
were defined on an ad-hoc basis. This approach made it difficult to
understand how the brand information was used across the Visual Studio
projects, and this in turn made it difficult to infer and maintain
relationships between the values.

Introduce a Visual Studio "Shared Project" to define all brand
identifiers and reference those definitions consistently.