Centralize installation logic & reduce admin reqs
Prior to this patch, the project's C++ code could only be built if
Microsoft Visual Studio was run with the privileges of a system
administrator.

Encapsulate installation steps in the MakeVoice.exe binary, and extend
that binary to require administrative privileges (allowing the code to
be built without them).

This change also supports forthcoming work to eliminate duplicated
installation logic between the project's Node.js code and C++ code.